// Firmenich.InternalSafetyDataSheet.Form/src/Interfaces/Item.ts
/* eslint-disable @typescript-eslint/naming-convention */
// Generated by https://quicktype.io

export interface IResponseItem {
    "odata.metadata"?: string;
    "odata.type"?: string;
    "odata.id"?: string;
    "odata.etag"?: string;
    "odata.editLink"?: string;
    FileSystemObjectType?: number;
    Id?: number;
    ServerRedirectedEmbedUri?: string;
    ServerRedirectedEmbedUrl?: string;
    ContentTypeId?: string;
    Title: string;
    ComplianceAssetId: undefined;
    CodeONU: string;
    Prodnom: string;
    Nomchim: string;
    Densvap: undefined;
    Dangers: string[];
    Densit_x00e9_: string;
    Poidsmol: string;
    NoCAS: string;
    Pointfus: string;
    Odeur: string[];
    Nomchimique: string;
    Pointeb: string;
    Stabilis: undefined;
    Pointinf: string;
    Pointec: string;
    OData__x0044_L50: string;
    Aspect: string[];
    Vmeppm: undefined;
    Tensvap: undefined;
    Solvants: string[];
    Vleppm: undefined;
    Olfact: string;
    Explos: string;
    Comment1: undefined;
    Environnement: undefined;
    Toxicit_x00e9_: string[];
    Protection: string[];
    Acciepand: string[];
    Accifeu: string[];
    Corpsecresp: string[];
    Corpsecpeau: string[];
    Corpsecyeux: string[];
    Refer: string[];
    Visadate: string;
    Datem1: string;
    Stated: string;
    Visapers: string;
    Visa: string;
    Comment: undefined;
    Datem: undefined;
    Modifdoc: undefined;
    Log: string;
    Dangerxi: string[];
    Pour_x0020_Approbation: boolean;
    Corrector_x0020_Decision: string;
    Archivage: undefined;
    Couleur: string;
    NotesID: string;
    DangersComplement: undefined;
    ToxiciteComplement: undefined;
    EpandageText: undefined;
    FeuText: undefined;
    InhalationText: undefined;
    PheuText: undefined;
    YeuxText: undefined;
    MesuresParticulieres: undefined;
    CL50inhal: undefined;
    DL50dermal: undefined;
    CE50mg: undefined;
    Infinflammation: string;
    Supinflammation: undefined;
    Mention_x0020_d_x2019_avertissem: string;
    AccifeuEdited: undefined;
    Notification: boolean;
    SQOPAM: number;
    Modified: string;
    ID: number;
    Created: string;
    AuthorId: number;
    EditorId: number;
    OData__UIVersionString: string;
    Attachments: boolean;
    GUID: string;
}